Bug fix for shrinking values added to resource environment.
What we really need is a way to temporarily pop the assertions to shrink the value in the root context, then restore them. There seems to be some stuff in the z3 API to do this (e.g. z3_solver_get_assertions), but it's not in the Haskell bindings and seems to be too subtle to attempt at the 11th hour.

So instead we have this new rather expensive shrink function that doesn't use the SAT solver.